requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.4
pydantic-core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
requests==2.31.0
python-dotenv==1.0.0
resend==0.11.1
